finalizando a controller estatistica

+++ Codigo/SeguradosAPI/Persistence/UsuarioHasPergunta.cs
@@ -4,7 +4,7 @@
     {
         public int IdUsuario { get; set; }
         public int IdPergunta { get; set; }
-        public byte Acertou { get; set; }
+        public int Acertou { get; set; }
 
         public Pergunta IdPerguntaNavigation { get; set; }
         public Usuario IdUsuarioNavigation { get; set; }
